황동찬 feat : 애니메이션 적용 및 테스트 레벨에 Player를 통한 확인용 액터 추가

+++ GameEngineContents/MapEditorWindow.cpp
@@ -518,7 +518,7 @@ void MapEditorWindow::ReadActor(std::shared_ptr<GameEngineLevel> Level)
 	std::vector<SponeMapActor> AllInfo = ProcessMapInfo::OpenFile(FilePath);
 	if (std::string::npos != FilePath.GetFileName().find("read.csv"))
 	{
-		int index = FilePath.GetFullPath().find("read.csv");
+		size_t index = FilePath.GetFullPath().find("read.csv");
 		FilePath.SetPath(FilePath.GetFullPath().substr(0, index));
 	}
 	for (SponeMapActor _str : AllInfo)
